requirements.txt
Flask==2.0.1
Flask_SocketIO==5.3.6
python-dotenv==1.0.1
